test: update tests for check restructures
