+++ tests/test_annotations.py
@@ -1,12 +1,12 @@
-"""Tests for annotations.misc check."""
+"""Tests for annotations check."""
 
-from proselint.checks.annotations import misc as chk
+from proselint.checks import annotations as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for annotations.misc."""
+    """The test class for annotations."""
 
     __test__ = True
 
@@ -16,7 +16,7 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for annotations.misc."""
+        """Basic smoke test for annotations."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert self.passes("""Add it to the to do list.""")
         assert not self.passes("""Add it to the TODO list.""")
