+++ tests/test_industrial_language_commercialese.py
@@ -1,12 +1,12 @@
-"""Tests for misc.commercialese check."""
+"""Tests for industrial_language.commercialese check."""
 
-from proselint.checks.misc import commercialese as chk
+from proselint.checks.industrial_language import commercialese as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for misc.commercialese."""
+    """The test class for industrial_language.commercialese."""
 
     __test__ = True
 
@@ -16,6 +16,6 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for misc.commercialese."""
+        """Basic smoke test for industrial_language.commercialese."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert not self.passes("""We regret to inform you of this.""")
