+++ tests/test_industrial_language_jargon.py
@@ -1,12 +1,12 @@
-"""Tests for jargon.misc check."""
+"""Tests for industrial_language.jargon check."""
 
-from proselint.checks.jargon import misc as chk
+from proselint.checks.industrial_language import jargon as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for jargon.misc."""
+    """The test class for industrial_language.jargon."""
 
     __test__ = True
 
@@ -16,6 +16,6 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for jargon.misc."""
+        """Basic smoke test for industrial_language.jargon."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert not self.passes("""I agree it's in the affirmative.""")
