+++ tests/test_malaproprisms.py
@@ -1,12 +1,12 @@
-"""Tests for malaproprisms.misc check."""
+"""Tests for malaproprisms check."""
 
-from proselint.checks.malapropisms import misc as chk
+from proselint.checks import malapropisms as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for malaproprisms.misc."""
+    """The test class for malaproprisms."""
 
     __test__ = True
 
@@ -16,6 +16,6 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for malaproprisms.misc."""
+        """Basic smoke test for malaproprisms."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert not self.passes("""Found in the Infinitesimal Universe.""")
