+++ tests/test_mondegreens.py
@@ -1,12 +1,12 @@
-"""Tests for mondegreens.misc check."""
+"""Tests for mondegreens check."""
 
-from proselint.checks.mondegreens import misc as chk
+from proselint.checks import mondegreens as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for mondegreens.misc."""
+    """The test class for mondegreens."""
 
     __test__ = True
 
@@ -16,7 +16,7 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for mondegreens.misc."""
+        """Basic smoke test for mondegreens."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert self.passes("""... and laid him on the green.""")
         assert not self.passes("""..and Lady Mondegreen.""")
