+++ tests/test_skunked_terms.py
@@ -1,12 +1,12 @@
-"""Tests for skunked_terms.misc check."""
+"""Tests for skunked_terms check."""
 
-from proselint.checks.skunked_terms import misc as chk
+from proselint.checks import skunked_terms as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for skunked_terms.misc."""
+    """The test class for skunked_terms."""
 
     __test__ = True
 
@@ -16,7 +16,7 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for skunked_terms.misc."""
+        """Basic smoke test for skunked_terms."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert not self.passes(
             """I gave an impassionate defence of the situation.""")
