+++ tests/test_social_awareness_nword.py
@@ -1,12 +1,12 @@
-"""Tests for cursing.nword check."""
+"""Tests for social_awareness.nword check."""
 
-from proselint.checks.cursing import nword as chk
+from proselint.checks.social_awareness import nword as chk
 
 from .check import Check
 
 
 class TestCheck(Check):
-    """The test class for cursing.nword."""
+    """The test class for social_awareness.nword."""
 
     __test__ = True
 
@@ -16,6 +16,6 @@ class TestCheck(Check):
         return chk
 
     def test_smoke(self):
-        """Basic smoke test for cursing.nword."""
+        """Basic smoke test for social_awareness.nword."""
         assert self.passes("""Smoke phrase with nothing flagged.""")
         assert not self.passes("""The n-word.""")
